Issue #16: Extracted structures to a new entity module.
